21-ICCV-P2PNet/requirements.txt
torch==1.5.0
torchvision==0.6.0
tensorboardX
easydict
pandas
scipy
matplotlib
Pillow
opencv-python